# backend/src/PromptTemplateManager.Application/DTOs/FolderResponse.cs
namespace PromptTemplateManager.Application.DTOs;

public class FolderResponse
{
    public Guid Id { get; set; }
    public string Name { get; set; } = string.Empty;
    public Guid? ParentFolderId { get; set; }
    public DateTime CreatedAt { get; set; }
    public DateTime UpdatedAt { get; set; }
}